% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterPlotMatrix.R
\name{getPlotConfig}
\alias{getPlotConfig}
\title{Asks to retrieve the plot configuration. 
Result will be sent through a reactive input.}
\usage{
getPlotConfig(id, configInputId)
}
\arguments{
\item{id}{Output variable to read from (id which references the requested plot).}

\item{configInputId}{Reactive input to write to.}
}
\value{
No return value, called from shiny applications for side effects.
}
\description{
Asks to retrieve the plot configuration. 
Result will be sent through a reactive input.
}
